+++ pkg/status/status.go
@@ -5,6 +5,9 @@ const (
 	// merge configmaps.
 	ConfigmapMergeFailedStatus = "configmap-merge-failed"
 
+	// DeletingStatus is set in the CR status while child apps are being deleted.
+	DeletingStatus = "deleting"
+
 	// ResourceNotFoundStatus is set in the CR status when there is an failure during
 	// finding dependents kubernete resources.
 	ResourceNotFoundStatus = "resource-not-found"