+++ service/controller/app/resource/childapps/create.go
@@ -1,0 +1,9 @@
+package childapps
+
+import (
+	"context"
+)
+
+func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
+	return nil
+}